exten prisma client to always return user items
