+++ packages/core/src/features/feed/actions.ts
@@ -17,22 +17,6 @@ export const getFeed = createAction(async ({ }, params) => {
       title: undefined
     },
     include: {
-      user: {
-        include: {
-          inventory: {
-            include: {
-              collection: {
-                where: {
-                  equipped: true
-                },
-                include: {
-                  item: true
-                }
-              }
-            }
-          }
-        }
-      },
       comments: true,
       category: true,
       tags: true
